db/requirements.txt
amqp==5.0.1
aniso8601==7.0.0
appdirs==1.4.4
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
asgiref==3.2.7
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.4.4
bcrypt==3.1.7
billiard==3.6.3.0
black==20.8b1
Brotli==1.0.9
celery==5.0.1
certifi==2020.4.5.1
cffi==1.14.0
channels==3.0.1
chardet==3.0.4
ChatterBot==1.0.1
chatterbot-corpus==1.2.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
codecov==2.1.10
collective.checkdocs==0.2
colorama==0.4.4
coloredlogs==10.0
constantly==15.1.0
coverage==5.3
crayons==0.4.0
cryptography==3.2.1
cycler==0.10.0
daphne==3.0.0
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.0.7
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-excel==0.0.10
django-filter==2.2.0
django-import-export==2.4.0
django-taggit==1.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
docutils==0.16
et-xmlfile==1.0.1
Faker==4.1.0
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Compress==1.8.0
fs==2.4.11
future==0.18.2
gease==0.0.11
graphene==2.1.8
graphene-django==2.9.1
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
html5lib==1.1
humanfriendly==8.2
hyperlink==20.0.1
idna==2.8
imtools==2.0.1
imutils==0.5.3
incremental==17.5.0
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jinja2-fsloader==0.3.0
joblib==0.17.0
kiwisolver==1.2.0
kombu==5.0.2
lml==0.1.0
lxml==4.6.1
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==2.2.5
mccabe==0.6.1
moban==0.8.2
moban-jinja2-github==0.0.4
mypy-extensions==0.4.3
nltk==3.5
nose==1.3.7
numpy==1.19.3
oauthlib==3.0.1
odfpy==1.4.1
opencv-python==4.4.0.46
openpyxl==3.0.5
packaging==18.0
pandas==1.1.4
pathspec==0.8.1
pep8==1.7.1
phonenumbers==8.12.6
Pillow==6.2.2
Pint==0.16.1
plotly==4.13.0
promise==2.3
prompt-toolkit==3.0.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyecharts-jupyter-installer==0.0.3
pyexcel==0.6.6
pyexcel-ezodf==0.3.4
pyexcel-handsontable==0.0.2
pyexcel-io==0.6.4
pyexcel-ods3==0.6.0
pyexcel-pygal==0.0.2
pyexcel-webio==0.1.4
pyexcel-xls==0.6.1
pyexcel-xlsx==0.6.0
pyflakes==2.2.0
pygal==2.4.0
Pygments==2.7.2
PyHamcrest==2.0.2
pymongo==3.11.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyreadline==2.1
python-bidi==0.4.2
python-dateutil==2.7.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==3.13
regex==2020.10.23
reportlab==3.5.55
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Rx==1.6.1
scipy==1.5.4
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.15.0
SQLAlchemy==1.2.19
sqlparse==0.3.1
stripe==2.27.0
tablib==2.0.0
text-unidecode==1.3
texttable==1.6.3
toml==0.10.2
tqdm==4.50.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson @ file:///C:/ci/ujson_1602507951182/work
urllib3==1.24.2
vine==5.0.0
virtualenv==20.0.34
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.1.2
